+++ src/matterexport/markdown.py
@@ -3,15 +3,17 @@ Exporter creating Markdown output
 """
 
 
+# pylint: disable=fixme
+# TODO: resolve todos and remove this pylint exception
+
+
 import sys
-import pprint
 
-from .exporter import Exporter
-from . import dataaccessor
 from matterbak import teams
+from .exporter import Exporter
 
 
-class Markdown_Exporter(Exporter):
+class MarkdownExporter(Exporter):
     """Exporter creating Markdown output"""
 
     def __init__(self, data_dir, output=None):
@@ -29,10 +31,11 @@ class Markdown_Exporter(Exporter):
             self.output = sys.stdout
         else:
             try:
-                self.output = open(output, 'w')
+                # pylint: disable=consider-using-with
+                self.output = open(output, 'w', encoding='utf8')
             except OSError as ex:
                 print(f"Cannot open {output} as output file: {ex}")
-                exit(1)
+                sys.exit(1)
 
     def close(self):
         """Close output file if applicable"""
@@ -51,17 +54,21 @@ class Markdown_Exporter(Exporter):
             heading: heading marker for title
         """
         timestamp = post.get_timestamp()
-        username = self.users.get_displayname(post.data['user_id'])
+        username = self.data_accessor.users.get_displayname(post.data['user_id'])
 
+        # TODO: Add option to not insert local image paths or to create a zip
+        #       file with Markdown and image files
         # TODO: get custom emojis from metadata?
         # TODO: replace in-text emoji names by images?
-        self._append_output(f"{heading} {username} at {timestamp.replace(microsecond=0)}\n{post.data['message']}\n")
+        self._append_output(f"{heading} {username} at"
+                            f" {timestamp.replace(microsecond=0)}\n{post.data['message']}\n")
 
         any_file = False
         for file_data, file_content_path in post.get_files():
             if not any_file:
                 self._append_output("Files:\n")
                 any_file = True
+            # TODO: In case of an image file append the image
             self._append_output(f"* {file_data['name']}: `{file_content_path}`")
         if any_file:
             self._append_output('')
@@ -71,7 +78,7 @@ class Markdown_Exporter(Exporter):
             if not any_reaction:
                 self._append_output("Reactions:\n")
                 any_reaction = True
-            username = self.users.get_displayname(reaction['user_id'])
+            username = self.data_accessor.users.get_displayname(reaction['user_id'])
             # TODO: insert emoji image
             self._append_output(f"* {username}: `{reaction['emoji_name']}`")
         if any_reaction:
@@ -82,35 +89,37 @@ class Markdown_Exporter(Exporter):
         channel_type = channel.metadata['type']
         match channel_type:
             case teams.CHANNEL_TYPE_DIRECT:
-                self._append_output(f"# Direct channel\n")
+                self._append_output("# Direct channel\n")
             case teams.CHANNEL_TYPE_GROUP:
-                self._append_output(f"# Group channel\n")
+                self._append_output("# Group channel\n")
             case _:
                 self._append_output(f"# {channel.get_name()}\n")
 
         self._append_output(f"{channel.metadata['header']}\n")
-        self._append_output(f"{channel.metadata['total_msg_count']} messages in {channel.metadata['total_msg_count_root']} threads or single messages\n")
+        self._append_output(f"{channel.metadata['total_msg_count']} messages in"
+                            f" {channel.metadata['total_msg_count_root']}"
+                            " threads or single messages\n")
 
         self._append_output("## Metadata\n")
-        creation_time = dataaccessor.get_timestamp(channel.metadata['create_at'])
+        creation_time = self.data_accessor.get_timestamp(channel.metadata['create_at'])
         self._append_output(f"Created at {creation_time.replace(microsecond=0)}\n")
 
         if channel_type in teams.CHANNEL_TYPE_TEAM:
             team_id = channel.metadata['team_id']
-            team = dataaccessor.Team(self.data_dir, team_id)
+            team = self.data_accessor.get_team(team_id)
             self._append_output(f"Belonging to team {team.get_name()}\n")
         else:
             self._append_output("### Members\n")
             for member in channel.members:
                 user_id = member['user_id']
-                user_name = self.users.get_displayname(user_id)
+                user_name = self.data_accessor.users.get_displayname(user_id)
                 self._append_output(f"* {user_name}")
             self._append_output("\n")
 
     def channel(self, identifier, by_threads=False):
         """Exporter for a channel, implementation for Exporter.channel()"""
 
-        channel = super().channel(identifier, by_threads)
+        channel = self.data_accessor.get_channel(identifier)
 
         self._append_metadata(channel)
 
@@ -121,8 +130,6 @@ class Markdown_Exporter(Exporter):
                 self._append_post(post)
         else:
             for thread in channel.get_threads():
-                self._append_output(f"### Thread")
+                self._append_output("### Thread")
                 for post in thread:
                     self._append_post(post, heading='####')
-
-